[FE] feat : TemplateListPage 라우팅 탭 구현
- TODO: uri에 따라 API 요청을 다르게 해서 적절하게 필터링된 데이터를 보여줘야 한다.
- TemplateStorePage -> TemplateListPage 네이밍 수정
