+++ frontend/src/service/@shared/layout/MainLayout/Header.tsx
@@ -39,7 +39,7 @@ function Header() {
 
         <ul className={styles.menuList}>
           <li className={styles.menuItem}>회고덕 소개</li>
-          <Link to={PAGE_LIST.TEMPLATE_STORE}>
+          <Link to={PAGE_LIST.TEMPLATE_LIST}>
             <li className={styles.menuItem}>템플릿 탐색</li>
           </Link>
           <li className={styles.menuItem}>타임라인</li>
